move modal to own file
